Combine attribute.rs attribute_set.rs
