+++ src/attr.rs
@@ -1,7 +1,13 @@
 use std::fmt;
 use std::borrow::Cow;
 
-use Attribute;
+use escape::Escaped;
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct Attribute<'a> {
+    pub name: Cow<'a, str>,
+    pub value: Cow<'a, str>,
+}
 
 #[derive(Clone, Debug)]
 pub struct AttributeSet<'a> {
@@ -18,9 +24,18 @@ impl<'a> AttributeSet<'a> {
     }
 }
 
-impl<'a, T> From<T> for AttributeSet<'a> where T: Into<Cow<'a, [Attribute<'a>]>> {
+impl<'a, T> From<T> for AttributeSet<'a>
+    where T: Into<Cow<'a, [Attribute<'a>]>>
+{
     fn from(attrs: T) -> AttributeSet<'a> {
         AttributeSet::new(attrs.into())
+    }
+}
+
+impl<'a> fmt::Display for Attribute<'a> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        // TODO handle invalid attribute names
+        write!(f, "{}=\"{}\"", self.name, Escaped(&self.value))
     }
 }
