+++ src/lib.rs
@@ -1,14 +1,12 @@
 #[macro_use]
 pub mod macros;
 
-mod attribute;
-mod attribute_set;
-
+mod attr;
 mod escape;
 mod event;
 mod stream;
 
 pub use event::Event;
-pub use attribute::Attribute;
-pub use attribute_set::AttributeSet;
+pub use attr::Attribute;
+pub use attr::AttributeSet;
 pub use stream::HtmlStreamer;
